# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pessoas",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/emailPessoas": {
            "get": {
                "tags": [
                    "EmailPessoas"
                ],
                "summary": "Lista de emails de pessoas",
                "description": "Retorna uma lista de emails de pessoas",
                "operationId": "getEmailPessoasList",
                "responses": {
                    "200": {
                        "description": "Sucesso. Lista de emails de pessoas retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a lista de emails de pessoas."
                    }
                }
            },
            "post": {
                "tags": [
                    "EmailPessoas"
                ],
                "summary": "Cadastro de email de pessoa",
                "description": "store",
                "operationId": "5bbed9cc9ef132fe216dd30c8042e17f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "emailPessoa",
                                    "idPessoa"
                                ],
                                "properties": {
                                    "emailPessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "leonnardo1588@gmail.com"
                                    },
                                    "idPessoa": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Email de pessoa cadastrado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível cadastrar o email de pessoa."
                    }
                }
            }
        },
        "/api/v1/emailPessoas/{id}": {
            "get": {
                "tags": [
                    "EmailPessoas"
                ],
                "summary": "Busca de email de pessoa pelo id",
                "description": "Retorna um email de pessoa pelo id",
                "operationId": "getEmailPessoaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do email de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Email de pessoa encontrado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível encontrar o email de pessoa."
                    }
                }
            },
            "put": {
                "tags": [
                    "EmailPessoas"
                ],
                "summary": "Atualização de email de pessoa",
                "description": "update",
                "operationId": "af50f612afe7902d27df603b3bcb6948",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do email de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "emailPessoa",
                                    "idPessoa"
                                ],
                                "properties": {
                                    "emailPessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "leonnardo1588@gmail.com"
                                    },
                                    "idPessoa": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Email de pessoa atualizado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível atualizar o email de pessoa."
                    }
                }
            },
            "delete": {
                "tags": [
                    "EmailPessoas"
                ],
                "summary": "Exclusão de email de pessoa",
                "description": "destroy",
                "operationId": "a7bad6845e56b23a78209761848f5ab6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do email de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Email de pessoa excluído com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível excluir o email de pessoa."
                    }
                }
            }
        },
...
        "/api/v1/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Mostra uma pessoa",
                "description": "Mostra uma pessoa",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Pessoa retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a pessoa."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Atualização de pessoa",
                "description": "update",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomePessoa",
                                    "dataNascimento",
                                    "cpf",
                                    "rg"
                                ],
                                "properties": {
                                    "nomePessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Leonardo"
                                    },
                                    "dataNascimento": {
                                        "description": "Data de nascimento",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF",
                                        "type": "string"
                                    },
                                    "rg": {
                                        "schema": "string",
                                        "description": "RG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Atualização realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a atualização."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Exclusão de pessoa",
                "description": "destroy",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Exclusão realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a exclusão."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "scheme": "http",
                "name": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}